// src/forgetrustee/frontend/src/types/generated/metadataFieldDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AllowedValuesRule } from './allowedValuesRule';
import type { MetadataFieldDtoAllowedValuesMode } from './metadataFieldDtoAllowedValuesMode';
import type { MetadataFieldDtoFields } from './metadataFieldDtoFields';
import type { MetadataFieldDtoType } from './metadataFieldDtoType';
import type { MetadataFieldOption } from './metadataFieldOption';
import type { RequiredRule } from './requiredRule';
import type { VisibilityRule } from './visibilityRule';

export interface MetadataFieldDto {
  /** The unique name of the field (machine-readable) */
  name: string;
  /** The human-readable label to show in the UI */
  label: string;
  heading?: string;
  placeholder?: string;
  tooltip?: string;
  helpText?: string;
  callToAction?: string;
  examples?: string[];
  minLength?: number;
  maxLength?: number;
  fields?: MetadataFieldDtoFields;
  /** The type of the field */
  type: MetadataFieldDtoType;
  multiple?: boolean;
  required?: boolean;
  requiredBy?: RequiredRule[];
  disabled?: boolean;
  readonly?: boolean;
  allowUnknown?: boolean;
  allowUnknownLabel?: string;
  allowNoAnswer?: boolean;
  allowNoAnswerLabel?: string;
  options?: MetadataFieldOption[];
  visible?: boolean;
  visibleBy?: VisibilityRule[];
  allowedValuesMode?: MetadataFieldDtoAllowedValuesMode;
  allowedValuesBy?: AllowedValuesRule[];
}